move a couple checks to a dedicated checks/tabular_glyphs.py file
* tnum_glyphs_equal_widths
* tabular_kerning
