+++ tests/test_checks_universal.py
@@ -1277,31 +1277,6 @@ def test_check_math_signs_width():
     assert_results_contain(check(font), WARN, "width-outliers")
 
 
-def test_check_tabular_kerning():
-    """Check tabular widths don't have kerning."""
-    check = CheckTester("tabular_kerning")
-
-    # Has no numerals at all
-    font = TEST_FILE("BadGrades/BadGrades-VF.ttf")
-    assert_SKIP(check(font))
-
-    # Has no tabular numerals
-    font = TEST_FILE("akshar/Akshar[wght].ttf")
-    assert_SKIP(check(font))
-
-    font = TEST_FILE("montserrat/Montserrat-Regular.ttf")
-    assert_PASS(check(font))
-
-    font = TEST_FILE("sharetech/ShareTech-Regular.ttf")
-    assert_results_contain(check(font), FAIL, "has-tabular-kerning")
-
-    # Ubuntu Sans has digraphs (like DZ) that get decomposed in ccmp
-    # and then have kerning between the individual D and Z, which
-    # used to throw off the check
-    font = TEST_FILE("ubuntusans/UbuntuSans[wdth,wght].ttf")
-    assert_PASS(check(font))
-
-
 def test_check_linegaps():
     """Checking Vertical Metric Linegaps."""
     check = CheckTester("linegaps")
